requirements.txt
Flask>=2.0.1
Werkzeug>=2.0.1
pandas
scikit-learn>=1.2.1
numpy
Cython>=0.29.24
mysql.connector
datetime